Adjustments to chatlog and message classes
