+++ discord_clone_rest_service/discord-api/src/main/java/discord_rest_api/models/DirectMessage.java
@@ -1,0 +1,14 @@
+package discord_rest_api.models;
+
+public class DirectMessage extends Message {
+    private int directChatId;
+
+    // Getters and Setters
+    public int getDirectChatId() {
+        return directChatId;
+    }
+
+    public void setDirectChatId(int directChatId) {
+        this.directChatId = directChatId;
+    }
+}
