+++ discord_clone_rest_service/discord-api/src/main/java/discord_rest_api/models/ServerMessage.java
@@ -1,0 +1,14 @@
+package discord_rest_api.models;
+
+public class ServerMessage extends Message {
+    private int channelId;
+
+    // Getters and Setters
+    public int getChannelId() {
+        return channelId;
+    }
+
+    public void setChannelId(int channelId) {
+        this.channelId = channelId;
+    }
+}
